requirements.txt
fastapi==0.115.12
pydantic==2.11.2
pymongo==4.11.3
uvicorn==0.34.0
python-dotenv==1.0.1

